make string escaping available in kore
